backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
google-cloud-aiplatform
Pillow